// FRONTEND/lib/types/database.ts
export type ProgrammingLanguage =
  | "CSHARP"
  | "NODE_JS"
  | "JAVA"
  | "GO"
  | "PYTHON"
  | "TYPESCRIPT"
  | "JAVASCRIPT"
  | "PHP"
  | "RUST"
  | "KOTLIN"
  | "SWIFT"
  | "OTHER";

export type PlatformType = "WEB" | "DESKTOP" | "MOBILE";